Rename composition/rhythm to composition/rhythmgen
